end of week 4.1

+++ week_2_exercises/ws1/light_travel.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
 int main() {
-    long long int light_speed = 299792458;
+    double light_speed = 299792458;
     int days;
 
     printf ("Please enter the number of days: ");
     scanf ("%d", &days);
 
-    long long int distance = days * 86400 * light_speed;
+    double distance = days * 86400 * light_speed;
 
-    printf ("%lld\n", distance);
+    printf ("%.15e\n", distance);
 
     return 0;
 }
